Resolves kjlubick/fb-contrib/issues/18
There is a bit of refactoring scattered throughout this.
I updated the message to show what was seen and what is expected, to
help identification of the problem.
The comments on lines 162 and 163 help explain what I do, basically when
we see a call to getName(), we pass the current uservalue forward
(passing the name of the class used to instantiate the logger)