Made fb-contrib a "OSGI bundle"
This makes it really easy for [the Eclipse
plugin](https://github.com/kjlubick/fb-contrib-eclipse-quick-fixes) to
access fb-contrib resources, like static lists, so as to avoid unecessary
duplication.

This does not impact how fb-contrib builds -- at all.

The changes to CharsetIssues were a test to see if exporting worked (and
also to allow the Eclipse plugin to access the map).

Inspired by:
http://stackoverflow.com/a/10881234/1447621