Integrate Travis CI without FindBugs
GitHub Issue #117.

Configure a basic Travis integration to build fb-contrib on Java 7 and 8
using Maven.

mvn install fails with a single fb-contrib error:

    Class com.mebigfatguy.fbcontrib.collect.CollectStatistics defines
    fields that are used only as locals
    ["com.mebigfatguy.fbcontrib.collect.CollectStatistics"] At
    CollectStatistics.java:[lines 41-129]

The warning is a true-positive that should probably be ignored. GitHub
Issue #124 tracks resolution.

Because Travis runs mvn install first [1], the build fails and mvn test
never gets to run. Override the default install directive with an
additional flag to skip FindBugs entirely, to be removed pending #124.

[1] https://docs.travis-ci.com/user/languages/java